requirements.txt
azure-functions==1.5.0
azure-common==1.1.25
azure-core==1.10.0
azure-nspkg==3.0.2
azure-storage-blob==12.7.1
certifi==2020.12.5
chardet==3.0.4
idna==2.8
joblib==1.0.0
numpy==1.19.5
pandas==1.2.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.22.0
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.25.11
